feat: make buildings and vills selectable
add selectable interface and add it to building and villager classes. Add script to child of villager to listen for mouseDown. The script needs to be on the same object as the collider. I do think I like this better than having the collider of the parent class though, as it is currently set for buildings.