fix: logic to allow for 1 prefab Warehouse
the first warehouse will be in the completed state and have resources. All other warehouses will start empty and in the placing state, this might not be a good solution if I want to add other buildings in the scene before the start of the game and initialise their state, but it works for now.