feat: add logic for consumption behaviour
Stamina decreases over time. When it is under 30 a consumption job is requested, which means a unit goes to a restaurant to eat and stamina is set back to 100.

Open issue is that when there is no consumption job available, the unit stops working. Because it is tricky to assign jobs asynchronously.