feat: make lumberjacks plant trees
IDestination was to big and building oriented. Had to remove some methods, it should really only be a location

create a LocationDestination to wrap a location in a GameObject, so that it can be used as destinatino

Add logic to plant trees is in a random location around the building.